Update docs
Let's leave `EncryptedRouter` in for now and bikeshed at the end. It
seems to me as appropriate as any other helper to include, but I'm open
to discussion. Also `KMS_CREDENTIALS` are not in use yet. I will test
with Azure and AWS prior to the end of this month.