Add kms credentials and providers env var tests
Ideally in schema.py instead of indexing KMS_CREDENTIALS with provider,
configure and use on-demand credentials. However, the implementation
in libmongocrypt appears to be that given a provider, the credentials
are acquired and used by PyMongo, which may not be suitable for use
in schema.py. However it may be possible to call a function in
libmongocrypt, instead of indexing KMS_CREDENTIALS with provider.